services/core/requirements.txt
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
prometheus-client==0.19.0
gunicorn==21.2.0
pyyaml==6.0.1